Renamed erebus/helper_functions.py into erebus/erebus_util.py; Added Permissions class which is created at runtime to load permissions to only have to read the json file one; Created exceptions.py with the InvalidAid exception
