+++ erebus/exceptions.py
@@ -1,0 +1,3 @@
+class InvalidAid(Exception):
+	def __init__(self, aid: str):
+		self.message = "The supplied aid {} was invalid".format(aid)
